graphql/notas_fiscais/requirements.txt
# Core da API
fastapi
strawberry-graphql[fastapi]
uvicorn[standard]

# Banco de dados
greenlet
oracledb
pydantic
sqlalchemy[asyncio]

# Configuração e ambiente
python-dotenv

# Logs, auditoria e observabilidade
ecs-logging==2.2.0

# Tipagem e lint
pylance
pylint
